loading the config data when adding the tab
The ExtJs tab will always go through the full life cycle including
rendering, even if the tab is never manually activated. To circumvent
the loading error when the repository is removed, loading of the repo
configuration is now done when adding the tab, instead of the previous
lazy-loading approach when rendering the tab.